--- a/.gitignore
+++ b/.gitignore
@@ -1,3 +1,4 @@
 node_modules
+lib/lexer.wat
 dist
 package-lock.json