--- a/src/lexer.h
+++ b/src/lexer.h
@@ -2,15 +2,15 @@
 #include <stdbool.h>
 #include <stddef.h>
 
-typedef unsigned char uchar_t;
-extern uchar_t __heap_base;
+typedef unsigned short char16_t;
+extern unsigned char __heap_base;
 
 struct Import {
   uint32_t start;
   uint32_t end;
-  uchar_t* endPos;
+  char16_t* endPos;
   uint32_t dynamic;
-  uchar_t* dynamicPos;
+  char16_t* dynamicPos;
   struct Import* next;
 };
 typedef struct Import Import;
@@ -25,8 +25,8 @@ typedef struct Export Export;
 // depth of template and brackets
 const uint32_t STACK_DEPTH = 1024;
 
-const uchar_t __empty_char = '\0';
-const uchar_t* EMPTY_CHAR = &__empty_char;
+const char16_t __empty_char = '\0';
+const char16_t* EMPTY_CHAR = &__empty_char;
 Import* first_import = NULL;
 Export* first_export = NULL;
 Import* import_read_head = NULL;
@@ -37,35 +37,31 @@ Export* export_write_head = NULL;
 void* analysis_base;
 void* analysis_head;
 
-size_t SOURCE_LEN;
-
 char templateStackDepth;
 char openTokenPosStackDepth;
 char templateDepth;
 char braceDepth;
-uchar_t* lastTokenPos;
-uchar_t* lastOpenTokenPos;
-uchar_t* pos;
-// js string index offset (accounting for different codepoint lengths)
-uchar_t* jsIndexOffset;
+char16_t* lastTokenPos;
+char16_t* lastOpenTokenPos;
+char16_t* pos;
 char* templateStack;
-uchar_t** openTokenPosStack;
+char16_t** openTokenPosStack;
 
 // Memory Structure:
-// allocateSource sets SOURCE_LEN
-// -> source [SOURCE_LEN]
-// -> analysis starts at source + SOURCE_LEN [bump allocated after source allocated]
-// Therefore: very important source* is last defined local!
+// -> source
+// -> analysis starts after source
 uint32_t parse_error;
 bool has_error = false;
-const uchar_t* source = &__heap_base;
+uint32_t sourceLen = 0;
+const char16_t* source = (void*)&__heap_base;
 
 void bail (uint32_t err);
 
 // allocateSource
-void* salloc (size_t len) {
-  SOURCE_LEN = len;
-  analysis_base = (void*)(source + len + 1 + ((len + 1) % 4));
+const char16_t* salloc (uint32_t utf16Len) {
+  sourceLen = utf16Len;
+  const char16_t* sourceEnd = source + utf16Len + 1;
+  analysis_base = (void*)sourceEnd;
   analysis_head = analysis_base;
   first_import = NULL;
   import_write_head = NULL;
@@ -73,10 +69,10 @@ void* salloc (size_t len) {
   first_export = NULL;
   export_write_head = NULL;
   export_read_head = NULL;
-  return (void*)source;
+  return source;
 }
 
-void addImport (uint32_t start, uint32_t end, uchar_t* endPos, uint32_t dynamic, uchar_t* dynamicPos) {
+void addImport (uint32_t start, uint32_t end, char16_t* endPos, uint32_t dynamic, char16_t* dynamicPos) {
   Import* import = (Import*)(analysis_head);
   analysis_head = analysis_head + sizeof(Import);
   if (import_write_head == NULL)
@@ -157,9 +153,9 @@ bool parse (uint32_t point);
 void tryParseImportStatement ();
 void tryParseExportStatement ();
 
-void readImportString (uchar_t ch);
+void readImportString (char16_t ch);
 
-uchar_t commentWhitespace ();
+char16_t commentWhitespace ();
 void singleQuoteString ();
 void doubleQuoteString ();
 void regularExpression ();
@@ -167,35 +163,35 @@ void templateString ();
 void blockComment ();
 void lineComment ();
 
-uchar_t readToWsOrPunctuator (uchar_t ch);
+char16_t readToWsOrPunctuator (char16_t ch);
 
-bool isBr (uchar_t c);
-bool isBrOrWs (uchar_t c);
-bool isBrOrWsOrPunctuator (uchar_t c);
-bool isBrOrWsOrPunctuatorNotDot (uchar_t c);
+bool isBr (char16_t c);
+bool isBrOrWs (char16_t c);
+bool isBrOrWsOrPunctuator (char16_t c);
+bool isBrOrWsOrPunctuatorNotDot (char16_t c);
 
-bool str_eq2 (uchar_t* pos, uchar_t c1, uchar_t c2);
-bool str_eq3 (uchar_t* pos, uchar_t c1, uchar_t c2, uchar_t c3);
-bool str_eq4 (uchar_t* pos, uchar_t c1, uchar_t c2, uchar_t c3, uchar_t c4);
-bool str_eq5 (uchar_t* pos, uchar_t c1, uchar_t c2, uchar_t c3, uchar_t c4, uchar_t c5);
-bool str_eq6 (uchar_t* pos, uchar_t c1, uchar_t c2, uchar_t c3, uchar_t c4, uchar_t c5, uchar_t c6);
-bool str_eq7 (uchar_t* pos, uchar_t c1, uchar_t c2, uchar_t c3, uchar_t c4, uchar_t c5, uchar_t c6, uchar_t c7);
+bool str_eq2 (char16_t* pos, char16_t c1, char16_t c2);
+bool str_eq3 (char16_t* pos, char16_t c1, char16_t c2, char16_t c3);
+bool str_eq4 (char16_t* pos, char16_t c1, char16_t c2, char16_t c3, char16_t c4);
+bool str_eq5 (char16_t* pos, char16_t c1, char16_t c2, char16_t c3, char16_t c4, char16_t c5);
+bool str_eq6 (char16_t* pos, char16_t c1, char16_t c2, char16_t c3, char16_t c4, char16_t c5, char16_t c6);
+bool str_eq7 (char16_t* pos, char16_t c1, char16_t c2, char16_t c3, char16_t c4, char16_t c5, char16_t c6, char16_t c7);
 
-bool readPrecedingKeyword2(uchar_t* pos, uchar_t c1, uchar_t c2);
-bool readPrecedingKeyword3(uchar_t* pos, uchar_t c1, uchar_t c2, uchar_t c3);
-bool readPrecedingKeyword4(uchar_t* pos, uchar_t c1, uchar_t c2, uchar_t c3, uchar_t c4);
-bool readPrecedingKeyword5(uchar_t* pos, uchar_t c1, uchar_t c2, uchar_t c3, uchar_t c4, uchar_t c5);
-bool readPrecedingKeyword6(uchar_t* pos, uchar_t c1, uchar_t c2, uchar_t c3, uchar_t c4, uchar_t c5, uchar_t c6);
-bool readPrecedingKeyword7(uchar_t* pos, uchar_t c1, uchar_t c2, uchar_t c3, uchar_t c4, uchar_t c5, uchar_t c6, uchar_t c7);
+bool readPrecedingKeyword2(char16_t* pos, char16_t c1, char16_t c2);
+bool readPrecedingKeyword3(char16_t* pos, char16_t c1, char16_t c2, char16_t c3);
+bool readPrecedingKeyword4(char16_t* pos, char16_t c1, char16_t c2, char16_t c3, char16_t c4);
+bool readPrecedingKeyword5(char16_t* pos, char16_t c1, char16_t c2, char16_t c3, char16_t c4, char16_t c5);
+bool readPrecedingKeyword6(char16_t* pos, char16_t c1, char16_t c2, char16_t c3, char16_t c4, char16_t c5, char16_t c6);
+bool readPrecedingKeyword7(char16_t* pos, char16_t c1, char16_t c2, char16_t c3, char16_t c4, char16_t c5, char16_t c6, char16_t c7);
 
-bool isExpressionKeyword (uchar_t* pos);
-bool isParenKeyword (uchar_t* pos);
-bool isPunctuator (uchar_t charCode);
-bool isExpressionPunctuator (uchar_t charCode);
-bool isExpressionTerminator (uchar_t* pos);
+bool isExpressionKeyword (char16_t* pos);
+bool isParenKeyword (char16_t* pos);
+bool isPunctuator (char16_t charCode);
+bool isExpressionPunctuator (char16_t charCode);
+bool isExpressionTerminator (char16_t* pos);
 
-void nextChar (uchar_t ch);
-void nextCharSurrogate (uchar_t ch);
-uchar_t readChar ();
+void nextChar (char16_t ch);
+void nextCharSurrogate (char16_t ch);
+char16_t readChar ();
 
 void syntaxError ();