… lib/lexer.wasm