This implements https://github.com/guybedford/es-module-lexer/issues/13, using UTF16 for the parser encoding as opposed to UTF8.

This reduces the amount of code (less surrogate handling) and removes the dependence on TextEncoder, which is a benefit as well.

This improves the performance considerably as detailed below.

Before this PR:

```
Module load time
> 6ms
Cold Run, All Samples
test/samples/*.js (2150 KiB)
> 29ms

Warm Runs (average of 25 runs)
test/samples/d3.js (491 KiB)
> 5.6ms
test/samples/d3.min.js (274 KiB)
> 3.44ms
test/samples/magic-string.js (34 KiB)
> 0.36ms
test/samples/magic-string.min.js (20 KiB)
> 0.04ms
test/samples/rollup.js (902 KiB)
> 9.24ms
test/samples/rollup.min.js (429 KiB)
> 5.24ms

Warm Runs, All Samples (average of 25 runs)
test/samples/*.js (2150 KiB)
> 24.8ms
```

After this PR:

```
Module load time
> 6ms
Cold Run, All Samples
test/samples/*.js (2150 KiB)
> 22ms

Warm Runs (average of 25 runs)
test/samples/d3.js (491 KiB)
> 4.2ms
test/samples/d3.min.js (274 KiB)
> 2.16ms
test/samples/magic-string.js (34 KiB)
> 0.04ms
test/samples/magic-string.min.js (20 KiB)
> 0ms
test/samples/rollup.js (902 KiB)
> 6.2ms
test/samples/rollup.min.js (429 KiB)
> 3.2ms

Warm Runs, All Samples (average of 25 runs)
test/samples/*.js (2150 KiB)
> 14.08ms
```